@@ -2,6 +2,7 @@ package ws
 
 import (
 	"encoding/json"
+
 	"github.com/gorilla/websocket"
 )
 
@@ -18,7 +19,7 @@ type wsUserStatusPayload struct {
 
 func (c *Client) readPump() {
 	defer func() {
-		c.room.unregister <- c
+		c.room.unregisterChan <- c
 		_ = c.conn.Close()
 	}()
 	c.conn.SetReadLimit(1 << 20)
@@ -42,18 +43,14 @@ func (c *Client) readPump() {
 			if err := json.Unmarshal(evt.Data, &payload); err != nil || payload.Username == "" {
 				continue
 			}
-			c.room.typingEventQ.add(payload.Username)
-			c.room.broadcastTypingQueue()
+			getHub().UpdateTyping(c.room.id, c, payload.Username, true)
 		case "stoppedTyping":
 			var payload wsUserStatusPayload
-			if err := json.Unmarshal(evt.Data, &payload); err != nil || payload.Username == "" {
-				continue
-			}
-			c.room.typingEventQ.remove(payload.Username)
-			c.room.broadcastTypingQueue()
+			_ = json.Unmarshal(evt.Data, &payload)
+			getHub().UpdateTyping(c.room.id, c, payload.Username, false)
 		case "joined", "left":
 			// Fan out ephemeral status events to everyone in the room.
-			c.room.broadcast <- data
+			c.room.broadcastChan <- data
 		default:
 			// Ignore other incoming event types for now.
 		}
